separate exitstatus.go
